intermediate push with more button changes: restoring wards to "schools" app.
It is not immediately obvious how to limit the schools dropdown menu because the schools.json file does not include ward number, grades, etc. That information is in the commute json. That said, I'll try quickly hacking transparent.js to read that information for all schools rather than just a cluster.